KAFKAC-66: Limit batches when writing to topic
Motivation
----------
On backfilling from large cluster, it is possible to get OOM,
when queue.poll() won't be empty for long enough time to
exhaust memory on the connector.

DCP flow control does not pause stream in this case, because
the ByteBuf have to be released right after transforming
into Kafka SourceRecord.

Modification
------------
Introduce new connector setting, which controls maximum number
of records in the batch for writing into Kafka topic. This number
only affect the case, when DCP snapshots are not used.

Result
------
The connector does not exhaust memory on large continuous DCP
transfers.

Change-Id: I7877d0eeb356b56303557baa7911e1e45f8cd05f
Reviewed-on: http://review.couchbase.org/74624
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>