KAFKAC-80 MutationToken Enabled in Kafka Connector to N1QL at plus
Motivation
----------

A Kafka consumer wants to run N1QL queries using a mutation token.
In order to make this possible, we need to include the vBucket UUID
and bucket name in the message so they can be used to construct
a mutation token.

Modifications
-------------

Unfortunately the existing Converter interface does not allow passing
in the vBucket UUID. Instead, this changes introduces a new
`SourceHandler` class to replace `Converter`. The new class is designed
with future extension in mind. It's an abstract class so we can
add methods as necessary. Instead of an argument list it uses a
parameter block so we can pass new arguments that custom implementations
can use (or ignore) without requiring breaking API changes.

The SourceHandler is no longer responsible for setting the
source partition and source offset. That's one less thing for
users to worry about (and there's no reason to customize it, anyway).

Users who don't need MutationTokens and who wish to continue working
with the deprecated Converter interface can do so; there's an adapter
that will make a Converter work like a SourceHandler.

Result
------

The info needed to build MutationTokens is present in Kafka messages.

No additional code changes are necessary to resolve KAFKAC-69
(Allow Source connector to split DCP stream and write into separate topics)

In the future it will be easier to modify the API without breaking it.

No compatibility issues are anticipated for existing users
(famous last words).

Change-Id: Ia248a13a0b1dc604f75cd92664ed28d8badb8c3e
Reviewed-on: http://review.couchbase.org/84360
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>