KAFKAC-85 Sink connector: Support deletion
Motivation
==========
Publishing a Kafka message with a null value is the idiomatic way
to delete a document. The sink connector should support this use case.

Modifications
=============
CouchbaseSinkTask.put checks if the record value is null. If so,
it deletes the Couchbase document whose ID matches the Kafka message key.

Incidental change: JsonBinaryDocument.create no longer throws NPE
when passed a null byte array as content. This is for consistency with
BinaryDocument, and has no effect on the behavior of the current code.

Result
======
The sink now interprets a null record value as a request to delete the
Couchbase document.

This change is not expected to negatively impact current users, since
prior to this change a null record value would cause a
NullPointerException.

Change-Id: Ib10bedbd3f660ac583899739fb110c786ce3c35e
Reviewed-on: http://review.couchbase.org/86732
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>