KAFKAC-84 Sink: Allow setting durability requirements for Couchbase writes
Motivation
==========
User would like to control the durability requirements observed by
the sink connector.

Modifications
=============
Add two new optional config properties to control persistence
and replication requirements:

  couchbase.durability.persist_to
  couchbase.durability.replicate_to

The possible values are the values of the PersistTo and ReplicateTo
enums. The default value for both is NONE (same as previous behavior),
so this change is not expected to impact current users.

Result
======
User can configure durability requirements. Connector terminates
if requirements cannot be met.

Change-Id: Ic3b5c32c48bf0de192d2cf39956c3ad77bcae608
Reviewed-on: http://review.couchbase.org/86854
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>