KAFKAC-133 Buffers not released on all code paths
Motivation
==========
Netty buffer leak detection shows some DCP mutation buffers
are not being released. This prompted a manual audit, which found that
buffers may leak when a connector task is restarted or under certain
error conditions.

Modifications
=============
Where appropriate, release buffers in 'catch' or 'finally' blocks.

Release buffers when the connector is stopped. (This happens during
workload redistribution when connector is running in distributed mode.)
Log a message "Releasing unconsumed events: N" where N is the number
of released buffers.

Modify the DCP control and data handlers to propagate exceptions
to the fatal error queue so that are reported to the Kafka connect
framework.

Result
======
Fewer opportunities for memory leaks.

More exceptions are propagated to the connect framework, which may
result in previously undetected failures being logged and causing the
connector to terminate rather than fail silently.

Change-Id: I39908d320b0e995db5d56fd3a7a3369d49416850
Reviewed-on: http://review.couchbase.org/103214
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>