KAFKAC-141 Add example of a custom Single Message Transform KAFKAC-142 Support publishing raw JSON without having to build example project
Motivation
==========
Users have expressed interest in modifying the document before writing
to Couchbase. Best way is with a Single Message Transform; we can help
by providing an example of how to write a custom transform.

Also, users have found it frustrating that it's necessary to compile
the example project in order to send documents as raw JSON. We should
support this out of the box.

Modifications
=============
Add an sample CustomTransform to the examples project.

Rename the examples project from `custom-source-handler` to
`custom-extensions` since now it has more than just a source handler.

Move the raw json source handler from the example project into
the core library. Avoid redundant de/serialization by directly
publishing byte arrays.

Add a DeserializeJson transform that allows other Single Message
Transforms to be used with the output of the RawJsonSourceHandler.
This is necessary because most other transforms are incompatible with
the byte array values produced my RawJsonSourceHandler.

Add a variant of RawJsonSourceHandler that includes Couchbase document
metadata in the message.

Add a DropIfNullValue transform that may be used to filter the
output of RawJsonSourceHandler to ignore deletion events.

Change-Id: Ib6d0f8a12f352c822cb4fab3f0be366bb797d0d8
Reviewed-on: http://review.couchbase.org/106156
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>