KAFKAC-154 Expose multi-network config options
Motivation
----------
By upgrading to the latest DCP client we got alternate address support
for free, but we still need to let users override the default strategy
(auto-detection) in case it fails for whatever reason.

Modifications
-------------
Add new "couchbase.network" config property (defaults to "auto").

Document it as "uncommitted", to be promoted to "committed" in next
minor release.

Change-Id: Ie94ba1b67c4c630c1fcc9be2801ca1ea790a2d21
Reviewed-on: http://review.couchbase.org/113535
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>