KAFKAC-194 Source: Support Kafka record headers
Motivation
----------
Headers are a great way to send document metadata separately
from the value of the Kafka record.

Customers have asked for header support in the past, and it will be
useful for communicating collection and scope IDs.

Modifications
-------------
Replace `CouchbaseSourceRecord` with `SourceRecordBuilder`.

Add builder method for setting headers.

Make `SourceHandler` an interface instead of an abstract class.

Change-Id: I4d3175a0af16809679abe4ea18b4b1bb585e43ba
Reviewed-on: http://review.couchbase.org/c/kafka-connect-couchbase/+/130277
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>