KAFKAC-195 Source: Publish to Kafka topic corresponding to collection
Motivation
----------
Support routing a document to the Kafka topic corresponding to the
document's scope and collection.

For example, if a document's scope is "tenant-foo" and its collection
is "invoices", setting the "couchbase.topic" config property to
"something.${scope}.${collection}" will result in the document
being published to the Kafka topic "something.tenant-foo.invoices".

Modifications
-------------
Treat "couchbase.topic" as a format string that recognizes placeholders
for the bucket, scope, and collection names.

Change-Id: Iae0a5293ce7add1013a25ce06c5ae59be284b807
Reviewed-on: http://review.couchbase.org/c/kafka-connect-couchbase/+/130364
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>