KAFKAC-157 Package connector for distribution on Confluent Hub
Motivation
----------
Packaging the connector in the Confluent component format
makes it easy to install the connector using the Confluent tooling.

It's also the required format for submission to the Confluent Hub
connector repository.

Developers who don't use Confluent can still use the connector by
unzipping the archive and installing the libraries manually.

Modifications
-------------
Use `kafka-connect-maven-plugin` to build the distribution archive
instead of the shade and assembly plugins.

Update the "how to release" guide to copy the archive from the
new location.

Change the project version from "dp.1" to "pre.1".

Change-Id: Ib4826ef6072c6422a951f0a1d1a9c6a9f54a991b
Reviewed-on: http://review.couchbase.org/c/kafka-connect-couchbase/+/130592
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>