KAFKAC-169 Sink: Guarantee documents are written in correct order
Motivation
----------
If a batch of sink records contains more than one version
of a document, it's theoretically possible for a stale version
to be written to Couchbase last, because documents in a batch
are written to Couchbase in parallel and asynchronously.

Modifications
-------------
Pre-process the batch to ignore all but the last version
of a document.

Change-Id: I7e7e54a1915a76878361ea84eeb8e6313b7f4d92
Reviewed-on: http://review.couchbase.org/c/kafka-connect-couchbase/+/131121
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>