KAFKAC-209 Source: Couchbase partition assignment creates unbalanced workloads
Motivation
----------
The partition assignment algorithm prioritizes minimizing the number
of DCP connections at the expense of equalizing the workload
among tasks. As a result, some users have seen dramatic workload
imbalances.

Another quirk of the algorithm is that it omitted partitions for which
there is no active partition. This could cause problems if the
connector starts when the Couchbase cluster topology is in flux.

Modifications
-------------
Assign partitions to tasks as evenly as possible. Each task
has always connected to every Couchbase server anyway,
so this shouldn't be any less efficient.

Borrow `ListHelper.chunks()` from the Elasticsearch connector
to split the partitions into evenly sized chunks.

Ignore bucket configs with zero partitions; these can appear
if the bucket was created within the past second or so.

Change-Id: I6ab12cc105bd6bda9e6d9b2152461550cf6593ed
Reviewed-on: http://review.couchbase.org/c/kafka-connect-couchbase/+/131483
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>