KAFKAC-220 Allow custom components to read connector config
Motivation
----------
Allow passing configuration options to a custom Filter or SourceHandler.

Modifications
-------------
Add an `init` method to the Filter and SourceHandler interfaces.
This method accepts a Map<String,String> representing the connector
config properties.

Also Considered
---------------
Could have passed the config Map to these components via an optional
constructor invoked via reflection.

The advantage of a constructor would be that config properties could be
read into `final` variables.

The constructor approach was not selected because it seemed more "magical"
and less discoverable than an interface method. Final variables are not
required for memory visibility in this case, since the Kafka Connect
framework uses a dedicated thread per task.

Change-Id: I1a3f1e7b1425d7dd50b7d98126cdb8b6348af05c
Reviewed-on: http://review.couchbase.org/c/kafka-connect-couchbase/+/134172
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>