KAFKAC-296 Source: Optionally publish ignored events to "black hole" topic
Add new `couchbase.black.hole.topic` source config property.

If this property is non-blank, the connector publishes a tiny
synthetic record to this topic whenever the Filter or SourceHandler
ignores a source event.

This lets the connector tell the Kafka Connect framework about the
source offset of the ignored event. Otherwise, a long sequence of
ignored events in a low-traffic deployment might cause the stored
source offset to lag too far behind the current source offset,
which can lead to rollbacks to zero when the connector is restarted.

Each record published to the black hole topic has one of 1024 possible
keys. Because the records themselves are not important, you can
minimize storage requirements by setting the black hole topic's
`cleanup.policy` to `compact`.

Change-Id: I5e4e1c954658bd478f13f0f3234c612630424d74
Reviewed-on: https://review.couchbase.org/c/kafka-connect-couchbase/+/179977
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>