KAFKAC-327 Sink: Support writing to multiple buckets
Motivation
----------
AnalyticsSinkHandler needs this feature, and it makes sense
for KV to implement it for config/feature parity.

Modifications
-------------
Add a new Keyspace class. It's a tuple of bucket, scope,
and collection. Bucket is nullable, but in practice is
only null with AnalyticsSinkHandler, and only when
the user omits the `couchbase.bucket` config property.

Update the sink connector to use Keyspace instead of
ScopeAndCollection. (Source handler still uses
ScopeAndCollection for now because it reinforces
the idea that the source reads from only one bucket.)

Sink config properties `couchbase.default.collection`
and `couchbase.topic.to.collection.map` now accept
an optional bucket prefix. If not specified, the bucket
defaults to `couchbase.bucket`.

Dots in bucket names can be escaped using SQL++ syntax
(enclose in backticks).

Modified AnalyticsSinkHandler to get the keyspace
(including bucket/database name) directly from
SinkHandlerParams.keyspace(), since a Keyspace object
knows how to format itself in a way that's compatible
with Analytics.

Credits
-------
This change is based on work originally done by Ritik Raj
for AnalyticsSinkHandler.

Change-Id: Ia5839f3298fcff8313b48c0b60fe6aecf8911ecb
Reviewed-on: https://review.couchbase.org/c/kafka-connect-couchbase/+/201482
Reviewed-by: Ritik Raj <raj.ritik9835@gmail.com>
Tested-by: Build Bot <build@couchbase.com>