KAFKAC-367 Support publishing multiple Kafka records for a single Couchbase event
Motivation
----------
Requested by users, and we've recently learned
how to do it safely.

Modifications
-------------
Add a `MultiSourceHandler` interface that is like
`SourceHandler` but allows returning multiple Kafka records.

Accept either a `SourceHandler` or a `MultiSourceHandler`.

If given a `SourceHandler`, wrap it in a `MultiSourceHandler`
so there's only one code path downstream.

In the Javadoc for `MultiSourceHandler`, document that this
is safe only if the Kafka Connect worker is configured with
`exactly.once.source.support=enabled`, which requires
Apache Kafka 3.3.0 or later.

Require JDK 9 when building the connector.
(JDK 8 has a bug that prevents it from working with
the JSpecify 'NullMarked' annotation.)

Change-Id: Ia25f5fccc671b36066549866f3c2eaeba6414182
Reviewed-on: https://review.couchbase.org/c/kafka-connect-couchbase/+/223707
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>