KAFKAC-401 Source: Events in system scope not actually filtered out by default
Motivation
----------
A bug prevented system scope filtering from working as advertised.

Modifications
-------------
Remove the CouchbaseSourceTask.filterIsNoop flag and the
optimization that skipped filtering if the flag was true.
Now the filter is always executed.

Updated the documentation to say the first connector version
with system scope filtering is 4.3.0 (instead of 4.2.4, when
the code was added but was ineffective).

Rejected Alternatives
---------------------
Could have updated the `filterIsNoop` check to look for
`AllPassIncludingSystemFilter` instead of `AllPassFilter`,
but that would have allowed similar bugs in the future.

Also, the optimization is less useful now that the default
filter is no longer a noop.

Change-Id: I842c1b9af88d8b7d1eee3b73bcb9abf31d2650f3
Reviewed-on: https://review.couchbase.org/c/kafka-connect-couchbase/+/232218
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>