Bump Avro dependency versions
Motivation
----------
The new `avro` dependency added for KAFKAC-399
brings in additional transitive dependencies.
Some of these dependency versions are on the
older side and have CVEs filed against them.

Modifications
-------------
Add dependency management entries for the
transitive dependencies of the `avro` library
so the current latest versions are used.

Update the avro library from 1.12.0 to 1.12.1.

Change-Id: I900b9072d40cd9cbc516266fa82db16d48c7f53b
Reviewed-on: https://review.couchbase.org/c/kafka-connect-couchbase/+/235533
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: <matthew.bray@couchbase.com>